requirements.txt
bcrypt==4.0.1
cryptography==42.0.5
fastapi==0.110.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
uvicorn==0.29.0
pdbpp==0.10.3
pytest==8.1.1
httpx==0.27.0